Replace with lalrpop
